add e2e dmf delete tests
Issue ZENKO-4179
