+++ tests/ctst/common/common.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import { tmpNameSync } from 'tmp';
-import { Given, setDefaultTimeout, Then } from '@cucumber/cucumber';
+import { Given, setDefaultTimeout, Then, When } from '@cucumber/cucumber';
 import { Constants, S3, Utils } from 'cli-testing';
 import Zenko from 'world/Zenko';
 import { extractPropertyFromResults, safeJsonParse } from './utils';
@@ -151,3 +151,93 @@ Then('object {string} should have the user metadata with key {string} and value 
         assert(head.Metadata[userMDKey]);
         assert(head.Metadata[userMDKey] === userMDValue);
     });
+
+// add a transition workflow to a bucket
+Given('a transition workflow to {string} location', async function (this: Zenko, location: string) {
+    this.resetCommand();
+    this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
+    this.addCommandParameter({
+        lifecycleConfiguration: JSON.stringify({
+            Rules: [
+                {
+                    Status: 'Enabled',
+                    Prefix: '',
+                    Transitions: [
+                        {
+                            Days: 1,
+                            StorageClass: location,
+                        },
+                    ],
+                },
+            ],
+        }),
+    });
+    await S3.putBucketLifecycleConfiguration(this.getCommandParameters());
+});
+
+When('i restore object {string} for {int} days', async function (this: Zenko, objectName: string, days: number) {
+    const objName = objectName ||  this.getSaved<string>('objectName');
+    this.resetCommand();
+    this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
+    this.addCommandParameter({ key: objName });
+    const versionId = this.getSaved<Map<string, string>>('createdObjects')?.get(objName);
+    if (versionId) {
+        this.addCommandParameter({ versionId });
+    }
+    this.addCommandParameter({ restoreRequest: `Days=${days}` });
+    await S3.restoreObject(this.getCommandParameters());
+});
+
+// wait for object to transition to a location or get restored from it
+Then('object {string} should be {string} and have the storage class {string}',
+    async function (this: Zenko, objectName: string, operation: string, storageClass: string) {
+        const objName = objectName ||  this.getSaved<string>('objectName');
+        this.resetCommand();
+        this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
+        this.addCommandParameter({ key: objName });
+        const versionId = this.getSaved<Map<string, string>>('createdObjects')?.get(objName);
+        if (versionId) {
+            this.addCommandParameter({ versionId });
+        }
+        let conditionOk = false;
+        while (!conditionOk) {
+            const res = await S3.headObject(this.getCommandParameters());
+            if (res.err) {
+                break;
+            }
+            assert(res.stdout);
+            const parsed = safeJsonParse(res.stdout);
+            assert(parsed.ok);
+            const head = parsed.result as {
+                StorageClass: string | undefined,
+                Restore: string | undefined,
+            };
+            const expectedClass = storageClass !== '' ? storageClass : undefined;
+            if (head?.StorageClass === expectedClass) {
+                conditionOk = true;
+            }
+            if (operation == 'restored') {
+                const isRestored = !!head?.Restore &&
+                    head.Restore.includes('ongoing-request="false", expiry-date=');
+                // if restore didn't get initiated fail immediately
+                const isPendingRestore = !!head?.Restore &&
+                    head.Restore.includes('ongoing-request="true"');
+                assert(isRestored || isPendingRestore, 'Restore didn\'t get initiated');
+                conditionOk = conditionOk && isRestored;
+            }
+            await Utils.sleep(3000);
+        }
+        assert(conditionOk);
+    });
+
+When('i delete object {string}', async function (this: Zenko, objectName: string) {
+    const objName = objectName ||  this.getSaved<string>('objectName');
+    this.resetCommand();
+    this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
+    this.addCommandParameter({ key: objName });
+    const versionId = this.getSaved<Map<string, string>>('createdObjects')?.get(objName);
+    if (versionId) {
+        this.addCommandParameter({ versionId });
+    }
+    await S3.deleteObject(this.getCommandParameters());
+});
